requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
groq
chromadb==0.4.4
python-dateutil==2.8.2
tqdm==4.66.1
lxml==4.9.3
sentence-transformers==2.2.2
streamlit==1.25.0
jsonschema>=4.17,<4.26
tenacity==8.2.2
